Refactor `ModuleManager` to be a part of the drawable stack
This change allows for all `Module`s to have a `ModuleContainer` which can manage each module on an individual level. It also allows for finer control over when data is loaded for each `Module`. This removes the need for the constructor that contained the `Storage` parameter. Also, sending and receiving OSC data is now centralised to the `ModuleManager`